chore: Moved ScanTarget and related functions into its own project
This is useful to have, and including SinglePassSigScanner just for ScanTarget is heavy.

This will be used in a future PR that adds runtime PDB support.
